feat: add thorough pre-merge test infrastructure
- tests/: doctest unit-test suite for wavefunction and element_data
  - tests/main.cpp: doctest entry point
  - tests/test_wavefunction.cpp: ~50 assertions covering associatedLaguerre,
    associatedLegendre, sphericalHarmonic (values, normalization, conjugate
    symmetry), radialWavefunction (exact R_10/R_21, normalization integral),
    wavefunction/probabilityDensity (origin values, spherical symmetry, 3D
    normalization), radialWavefunctionZ/wavefunctionZ (Z-scaling, Z=1 parity),
    computeVolumeData (size, range, all-zero for invalid QN),
    suggestedHalfSize (formula, monotonicity)
  - tests/test_element_data.cpp: ~35 assertions covering findElementByZ,
    findElementBySymbol, ELEMENTS data integrity (Z/index, Z_eff, outermost_n/l,
    periods), NUM_ELEMENTS, cross-lookup consistency
  - tests/CMakeLists.txt: FetchContent doctest v2.4.11, CTest registration
    with physics;element_data labels
- CMakeLists.txt: add BUILD_TESTS option (default ON), add_subdirectory(tests)
- .github/workflows/build.yml: add `test` job (ctest after linux build) and
  `lint` job (clang-tidy + cppcheck on wavefunction/element_data)
- .github/pull_request_template.md: expand Testing checklist with 4 new items
- .clang-tidy: enable bugprone-*/performance-*/readability-* with targeted
  disables for physics-code conventions; WarningsAsErrors='*'
- src/wavefunction.cpp: fix implicit int widening in computeVolumeData resize
  (detected by bugprone-implicit-widening-of-multiplication-result)

+++ src/wavefunction.cpp
@@ -243,7 +243,7 @@ void computeVolumeData(int n, int l, int m,
                         int resolution, double halfSize,
                         std::vector<float> &data)
 {
-    data.resize(resolution * resolution * resolution);
+    data.resize(static_cast<size_t>(resolution) * resolution * resolution);
 
     double dx = 2.0 * halfSize / (resolution - 1);
